+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SDataTable` component now supports `sortDescFirst` prop to set the default sort direction.
 
 ## 2.0.0-beta.45 - 2023-12-05